% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunction.R
\name{bootpval}
\alias{bootpval}
\title{Bootstrapped P-Value for Single Sample T-Test}
\usage{
bootpval(x, mu0 = 0, conf.level = 0.95, iter = 3000, test = "two")
}
\arguments{
\item{x}{A numeric vector representing the sample data.}

\item{mu0}{The null hypothesis mean value (default is 0).}

\item{conf.level}{Confidence level for the interval (default is 0.95).}

\item{iter}{Number of bootstrap iterations (default is 3000).}

\item{test}{The type of test ("two", "upper", or "lower").}
}
\value{
A list containing the p-value, test statistic, confidence interval, and additional details about the test.
}
\description{
Performs a bootstrapped hypothesis test for a single sample t-test, providing the p-value, confidence interval, and test statistic.
}
\examples{
set.seed(55)
x1 <- rnorm(30, mean = 25, sd = 5)
bootpval(x = x1, mu0 = 23, test = "two")
}
