% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
A data frame with X rows and Y columns:
\describe{
  \item{column1}{Description of column1}
  \item{column2}{Description of column2}
  ...
}
}
\source{
<source of the data, if applicable>
}
\usage{
fire
}
\description{
A dataset containing information about fire damage incidents.
}
\keyword{datasets}
