% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Confidence Interval Function}
\usage{
myci(x, alpha = 0.05)
}
\arguments{
\item{x}{A numeric vector representing the sample data.}

\item{alpha}{The significance level for the confidence interval (default is 0.05 for a 95% CI).}
}
\value{
A named vector with the lower and upper bounds of the confidence interval.
}
\description{
Confidence Interval Function
}
