% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L8function.R
\name{myclt}
\alias{myclt}
\title{Central Limit Theorem Simulation for Uniform Distribution}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{Sample size for each iteration.}

\item{iter}{Number of iterations.}
}
\value{
A histogram of the sums of samples from the uniform distribution.
}
\description{
This function simulates the Central Limit Theorem (CLT) by sampling from a
uniform distribution and calculating the sum of the samples. It then creates
a histogram of the sums.
}
