% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood Estimation Function}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{Likelihood function to be evaluated}

\item{x}{Vector of data points}

\item{param}{Vector of parameter values}

\item{...}{Additional arguments for plot customization}
}
\value{
List containing the index, parameter value, log-likelihood value, and slope values at max
}
\description{
This function computes the maximum likelihood estimate for a given likelihood function.
}
