% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Calculate the Number of Tickets to Sell}
\usage{
ntickets(N, gamma, p)
}
\arguments{
\item{N}{Integer. Number of seats in the flight.}

\item{gamma}{Numeric. Probability that the flight will be overbooked.}

\item{p}{Numeric. Probability of a "show".}
}
\value{
A named list containing:
\describe{
  \item{nd}{Number of tickets calculated using the discrete distribution.}
  \item{nc}{Number of tickets calculated using the normal approximation.}
  \item{N}{Number of seats in the flight.}
  \item{p}{Probability of a "show".}
  \item{gamma}{Probability that the flight will be overbooked.}
}
}
\description{
This function calculates the number of tickets to sell for a flight given the number of seats, the probability of a "show", and the probability that the flight will be overbooked.
}
\examples{
ntickets(N = 400, gamma = 0.02, p = 0.95)
}
